Deleted saving utilities
