+++ src/tensor_parallel/__init__.py
@@ -3,7 +3,7 @@ A prototype TensorParallel wrapper that works without torch.distributed.
 Splits linear, conv and some other layers between GPUs
 """
 
-from tensor_parallel.accelerate import load_separate_shards, save_separate_shards
+from tensor_parallel.dispatch import load_and_dispatch_separate_shards
 from tensor_parallel.factory import tensor_parallel
 from tensor_parallel.pretrained_model import TensorParallelPreTrainedModel
 from tensor_parallel.sharding import Sharded
