+++ tests/test_saving.py
@@ -8,10 +8,11 @@ from tensor_parallel import (
     Sharded,
     TensorParallel,
     TensorParallelPreTrainedModel,
-    load_separate_shards,
-    save_separate_shards,
+    load_and_dispatch_separate_shards,
     tensor_parallel,
 )
+
+PATH_TO_SAVE = "/tmp/"
 
 
 @pytest.mark.parametrize("devices", [("cpu",) * 2, ("cpu",) * 3])
@@ -94,10 +95,12 @@ def test_save_shards_load_shards(devices, model_name, shraded_class):
     model = BertModel.from_pretrained(model_name).to(devices[0])
     model_tp = shraded_class(model, devices)
 
-    save_separate_shards(model_tp, "/tmp")
+    model_tp.set_preserve_shards_when_saving(True)
+    torch.save(model_tp.state_dict(), PATH_TO_SAVE + "test_save_shards_load_shards.bin")
     del model_tp
 
     with init_empty_weights():
         model_tp = shraded_class(BertModel.from_pretrained(model_name), ("meta",) * len(devices))
 
-    load_separate_shards(model_tp, "/tmp/index.json", devices=devices)
+    load_and_dispatch_separate_shards(model_tp, PATH_TO_SAVE + "test_save_shards_load_shards.bin", device_ids=devices)
+    assert not "meta" in [p.device.type for p in model_tp.parameters()]
